requirements.txt
Flask==2.3.3
python-dotenv==1.0.0
pyairtable==2.3.0
Flask-Cors
requests
gunicorn